biopb-base/requirements.txt
biopb==0.2.0
grpcio-health-checking
imageio
typer